* Disable hash in version * Change new page default title and content * Admin: Users and Groups action now have approve modal * Profile: revoke token now have approve modal

+++ src/Mimisbrunnr.Web.Host/ClientApp/src/components/people/profile/settings/ApiTokens.vue
@@ -60,6 +60,16 @@ export default {
       this.tokens = tokenRequest.data;
     },
     revoke: async function (token) {
+      var approve = await this.$bvModal.msgBoxConfirm(this.$t("profile.settings.tokens.approveModal.revoke"), {
+            title: this.$t("profile.settings.tokens.approveModal.title"),
+            centered: true,
+            size: 'sm',
+            buttonSize: 'sm',
+            cancelTitle: this.$t("profile.settings.tokens.approveModal.cancel"),
+            okTitle: this.$t("profile.settings.tokens.approveModal.ok"),
+            okVariant: 'danger',
+      });
+      if(!approve) return;
       await axios.delete(`/api/account/token/${token.id}`);
       await this.load();
     },
