+++ src/Mimisbrunnr.Web.Host/ClientApp/src/views/Admin/Groups.vue
@@ -42,7 +42,7 @@
         {{ $t("admin.groups.loadMore") }}
       </b-button>
     </b-card>
-    <group-modal />
+    <group-modal :createAction="loadGroups" />
   </b-container>
 </template>
 
@@ -97,7 +97,19 @@ export default {
       this.loading = false;
     },
     removeGroup: async function (group) {
-      await GroupService.deleteGroup(group);
+        var approve = await this.$bvModal.msgBoxConfirm(this.$t("admin.groups.approveModal.deleteGroup"), {
+            title: this.$t("admin.groups.approveModal.title"),
+            centered: true,
+            size: 'sm',
+            buttonSize: 'sm',
+            cancelTitle: this.$t("admin.groups.approveModal.cancel"),
+            okTitle: this.$t("admin.groups.approveModal.ok"),
+            okVariant: 'danger',
+        });
+        if(!approve) return;
+        await GroupService.deleteGroup(group);
+        this.groups = [];
+        this.loadGroups();
     },
   },
   mounted() {
